Add face recognition and license plate recognition to settings UI (#17152)
* Refactor explore settings to classification settings

* Cleanup

* Add face config section

* Add license plate recognition to settings

* Update face recognition docs

* Fix variable usage

* Fix typo

* Update docs/docs/configuration/face_recognition.md


* Improve spacing and add face library to mobile

* Clarify docs

---------
